+++ src/components/Layout/Header/Brandlogo.tsx
@@ -10,7 +10,6 @@ export const BrandLogo: FC<BrandLogoProps> = ({ logoMode }) => {
     <Link to="/" className="logo">
       <img
         src={logoMode === "light" ? "/images/logo-dark.png" : "/images/logo-white.png"}
-        className={logoMode === "light" ? "light-logo" : "dark-logo"}
         alt="Zartek Technologies"
       />
     </Link>